% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccmppR}
\alias{ccmppR}
\title{Simulate cohort component population projection}
\usage{
ccmppR(basepop, sx, fx, srb, age_span, fx_idx)
}
\arguments{
\item{basepop}{vector of baseline population size.}

\item{sx}{a matrix of survivorship probabilities with column for each
projection step.}

\item{fx}{a matrix of fertility rates, only amongst fertile ages, with
column for each projection step.}

\item{srb}{a vector of sex ratio at birth.}

\item{age_span}{the interval for age and projection time step.}

\item{fx_idx}{first in}
}
\value{
A matrix of projected population at each step
}
\description{
Simulate cohort component population projection
}
\examples{

library(popReconstruct)
data(burkina_faso_females)
ccmppR(basepop = as.numeric(burkina.faso.females$baseline.pop.counts),
       sx = burkina.faso.females$survival.proportions,
       fx = burkina.faso.females$fertility.rates[4:10, 1],
       srb = rep(1.05, ncol(burkina.faso.females$survival.proportions)),
       age_span = 5,
       fx_idx = 4)
}
