% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{fit_tmb}
\alias{fit_tmb}
\title{Fit TMB model}
\usage{
fit_tmb(tmb_input, outer_verbose = TRUE, inner_verbose = FALSE, max_iter = 250)
}
\arguments{
\item{tmb_input}{A list containing `data` and `par_init`.}

\item{outer_verbose}{If `TRUE` print function and parameters every iteration.}

\item{inner_verbose}{If `TRUE` print TMB inner optimization tracing information.}

\item{max_iter}{Maximum number of iterations.}
}
\value{
Fit model object.
}
\description{
Fit TMB model
}
