% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_leslie_matrixR}
\alias{make_leslie_matrixR}
\title{Construct a sparse Leslie matrix}
\usage{
make_leslie_matrixR(sx, fx, srb, age_span, fx_idx)
}
\arguments{
\item{sx}{a vector of survivorship probabilities.}

\item{fx}{a vector of fertility rates, only amongst fertile ages.}

\item{srb}{sex ratio at birth.}

\item{age_span}{the interval for age and projection time step.}

\item{fx_idx}{first in}
}
\description{
Construct a sparse Leslie matrix
}
\details{
The first index of fertility `fx_idx` must be greater than 1, that is, it is
assumed there is no fertility in the youngest age group. The reason for this
restriction is that fertility among the youngest age group would require 
calculation of offspring among births during the second half of the
projection interval.
}
\examples{

library(popReconstruct)
data(burkina_faso_females)
make_leslie_matrixR(sx = burkina.faso.females$survival.proportions[,1],
                    fx = burkina.faso.females$fertility.rates[4:10, 1],
                    srb = 1.05,
                    age_span = 5,
                    fx_idx = 4)

}
