% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{make_tmb_obj}
\alias{make_tmb_obj}
\title{Make TMB objective function object}
\usage{
make_tmb_obj(
  data,
  par,
  model = "ccmpp_tmb",
  inner_verbose = FALSE,
  calc_outputs = TRUE
)
}
\arguments{
\item{data}{List of TMB data inputs.}

\item{par}{List of initial parameters for TMB model.}

\item{model}{Name of TMB model to be used.}

\item{inner_verbose}{Logical flag whether to print TMB inner optimization
tracing information.}

\item{calc_outputs}{Logical flag to be passed to model object whether to
calculate outputs not required for objective function.}
}
\value{
TMB objective function object with class `tmb_obj`.
}
\description{
Make TMB objective function object
}
\seealso{
This returns the object created by [TMB::MakeADFun].
}
