% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-frames.R
\name{proj_to_long}
\alias{proj_to_long}
\title{Convert projection output to long format}
\usage{
proj_to_long(proj, mf, value_col = "value")
}
\arguments{
\item{proj}{CCMPP projection output from [`ccmppR()`].}

\item{mf}{model frames list from [`projection_model_frames()`].}

\item{value_col}{column name for projection output.}
}
\description{
Convert projection output to long format
}
