% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-frames.R
\name{projection_indices}
\alias{projection_indices}
\alias{projection_model_frames}
\title{Construct indices for population projection}
\usage{
projection_indices(
  period_start,
  interval,
  n_periods,
  n_ages,
  fx_idx,
  n_fx,
  n_sexes = 1
)

projection_model_frames(indices)
}
\arguments{
\item{period_start}{start of first projection period.}

\item{interval}{projection interval (e.g. 1 year or 5 year).}

\item{n_periods}{number of projection periods.}

\item{n_ages}{number of age groups.}

\item{fx_idx}{first age index exposed to fertility.}

\item{n_fx}{number of age groups exposed to fertility.}

\item{indices}{list of indices, from [projection_indices()].}

\item{sexes}{sexes included in population projection.}
}
\description{
Construct indices for population projection
}
\examples{

indices5 <- projection_indices(period_start = 1950,
                              interval = 5,
                              n_periods = 12,
                              n_ages = 19,
                              fx_idx = 4,
                              n_fx = 7,
                              n_sexes = 1)

}
