% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb-model.R
\name{sample_tmb}
\alias{sample_tmb}
\title{Sample TMB fit}
\usage{
sample_tmb(
  fit,
  nsample = 1000,
  rng_seed = NULL,
  random_only = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{fit}{The TMB fit from [fit_tmb()].}

\item{nsample}{Number of samples to draw.}

\item{rng_seed}{Seed passed to [set.seed()].}

\item{random_only}{Logigcal whether to sample only random effects conditional
on optimized fixed effect values.}

\item{verbose}{If TRUE prints additional information.}
}
\value{
Fit object with additional list element `sample`.
}
\description{
Sample TMB fit
}
