% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ccmppR}
\alias{ccmppR}
\alias{ccmpp_leslieR}
\title{Simulate cohort component population projection}
\usage{
ccmppR(basepop, sx, fx, gx, srb, age_span, fx_idx)

ccmpp_leslieR(basepop, sx, fx, gx, srb, age_span, fx_idx)
}
\arguments{
\item{basepop}{vector of baseline population size.}

\item{sx}{a matrix of survivorship probabilities.}

\item{fx}{a matrix of fertility rates, only amongst fertile ages.}

\item{gx}{a matrix of proportion of migrants during projection period.}

\item{srb}{a vector of sex ratio at birth.}

\item{age_span}{the interval for age and projection time step.}

\item{fx_idx}{first in}
}
\value{
A matrix of projected population at each step
}
\description{
Simulate cohort component population projection
}
\details{
Arguments `sx`, `fx`, and `gx` are matrices with one column for each
projection period, and `srb` is a vector of length number of projection 
periods.
}
\examples{

library(popReconstruct)
data(burkina_faso_females)
ccmpp_leslieR(basepop = as.numeric(burkina.faso.females$baseline.pop.counts),
              sx = burkina.faso.females$survival.proportions,
              fx = burkina.faso.females$fertility.rates[4:10, 1],
              gx = burkina.faso.females$migration.proportions,
              srb = rep(1.05, ncol(burkina.faso.females$survival.proportions)),
              age_span = 5,
              fx_idx = 4)
}
